+++ api/v1/urls.py
@@ -1,7 +1,7 @@
 from django.urls import path, include
 
 urlpatterns = [
-  path("",include('apps.restaurant.urls'))
+  path("",include('apps.restaurant.urls')),
+  path("",include('apps.order.urls')),
     # path('flight/', include(('apps.flightapp.urls', 'flightapp'), namespace='flightapp')),
-    # path('route/', include(('apps.route.urls', 'route'), namespace='route')),
 ]
