+++ apps/order/readme.md
@@ -1,0 +1,2 @@
+# Order App
+> This app contains the Payment and delivery method , Cart, CartItem, Order and OrderItem and their respective model, admin, serializer, viewsets, urls, permissions
