+++ apps/restaurant/admin.py
@@ -1,17 +1,10 @@
 from django.contrib import admin
-from django.db.models.aggregates import Count
 from .models import (
-    Address, Cuisine, Restaurant, Dish, FoodCategory, Cart, CartItem, PaymentMethod, Order, OrderItem,
-    DeliveryMethod
-
+    Address, Cuisine, Restaurant, Dish, FoodCategory,
     )
 
 # Register your models here.
-
-
 admin.site.register(Address)
-admin.site.register(PaymentMethod)
-admin.site.register(DeliveryMethod)
 
 
 @admin.register(Cuisine)
@@ -32,15 +25,15 @@ class CuisineAdmin(admin.ModelAdmin):
 class FoodCategoryAdmin(admin.ModelAdmin):
     prepopulated_fields = {'slug': ['name']}
     fields = ('name', 'slug', 'description')
-    list_display = ('id', 'name', 'slug', 'dish_count')
+    list_display = ('id', 'name', 'slug', 'menu_count')
     list_editable = ('name',)
     list_filter = ('name',)
     search_fields = ('name', 'description')
     show_facets = admin.ShowFacets.ALWAYS
 
-    @admin.display(ordering='dish_count')
-    def dish_count(self, obj):
-        return obj.dish_count
+    @admin.display(ordering='menu_count')
+    def menu_count(self, obj):
+        return obj.menu_count
 
 
 @admin.register(Restaurant)
@@ -80,44 +73,3 @@ class DishAdmin(admin.ModelAdmin):
     def display_name(self, dish):
         return dish.name
 
-
-admin.site.register(Cart)
-admin.site.register(CartItem)
-
-
-class OrderItemInline(admin.TabularInline):
-    model = OrderItem
-    extra = 0
-    fields = ('dish_name', 'unit_price', 'quantity')
-
-
-@admin.register(Order)
-class OrderAdmin(admin.ModelAdmin):
-    list_display = ('order_number', 'customer', 'status',
-                    'payment_status', 'total', 'created_at')
-    list_filter = ('status', 'payment_status', 'created_at')
-    search_fields = ('order_number', 'customer__email')
-    readonly_fields = ('order_number', 'subtotal',
-                       'total', 'created_at', 'updated_at')
-    fieldsets = (
-        (None, {'fields': ('order_number', 'customer', 'status', 'payment_status')
-                }),
-        ('Financials', {
-            'fields': ('subtotal', 'delivery_fee', 'tax', 'total')
-        }),
-        ('Delivery', {
-            'fields': ('delivery_address', 'special_instructions', 'delivered_at')
-        }),
-    )
-    inlines = [OrderItemInline]
-
-
-@admin.register(OrderItem)
-class OrderItemAdmin(admin.ModelAdmin):
-    list_display = ('order', 'dish_name','unit_price' ,'quantity','sub_total')
-    list_select_related = ('order',)
-    readonly_fields = ('unit_price','quantity')
-    search_fields = ('order__order_number', 'dish_name')
-
-    def sub_total(self, obj):
-        return obj.unit_price * obj.quantity
