+++ apps/restaurant/managers.py
@@ -1,6 +1,7 @@
 from django.db import models
 from django.db.models.aggregates import Count
 
+
 class CuisineManager(models.Manager):
     def get_queryset(self):
         return super().get_queryset().annotate(restaurant_count=Count('restaurants'))
@@ -8,9 +9,9 @@ class CuisineManager(models.Manager):
 
 class FoodCategoryManager(models.Manager):
     def get_queryset(self):
-        return super().get_queryset().annotate(dish_count=Count('dishes'))
+        return super().get_queryset().annotate(menu_count=Count('dishes'))
 
 
 class RestaurantManager(models.Manager):
     def get_queryset(self):
-        return super().get_queryset().annotate(dish_count=Count('dishes'))
+        return super().get_queryset().annotate(menu_count=Count('dishes'))
