+++ config/settings/base.py
@@ -48,6 +48,7 @@ INSTALLED_APPS = [
     'api',
     'apps.authentication',
     'apps.restaurant',
+    'apps.order',
 ]
